requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
joblib
pandas
scikit-learn
lightgbm
python-dotenv